refactor: 認証の共通基盤を用途ディレクトリと同軸にせず src 直下へ据える
supabase-client / supabase-result は各クライアントが土台とする全体の基盤であり、
oauth / passkey などの用途とは軸が異なる。用途ディレクトリの兄弟に置くと粒度が
混ざって読みづらいため、errors と同じく src 直下の基盤モジュールとして据え置く。
